scripts/TapoLight/pyproject.toml
[project]
name = "tapolight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "tapo>=0.8.11",
]
